requirements.txt
Pillow==8.4.0
tqdm==4.24.0
